Fix issue with advanced search to pick the correct enrollment
This commit addresses following two issues,

1. Advanced search feature returns the oldest enrollment instance

This issue occurs when the same device is enrolled for multiple users. The cause of the issue is filtering the devices by device identifier which is a unique property for a device. This commit fixes the issue by filtering the devices from enrollment id and owner instead of the device identifier.

2. All devices do not return for privileged users when searched through advanced-search

The cause of the issue is not considering the user permissions when filtering the devices. This commit fixes the issue by validating the permissions of the logged in user before filtering.

Fixes entgra/product-iots#13
Fixes entgra/product-iots#14