Fix the task running issue in tenant mode
If a run time error encountered while sending a notification(e.g:- Push Notification ) to a device when the task is running for tenants, it terminates the task running and notifications won't be sending to devices which are enrolled to other tenants as well. Hence cath the exception in sendNotification method and log.

Further, since this is critical path, added few debug logs as well.