Fix app installing issue for same package name
If we add same installer as a new app  by giving different name, it was able to add as new application. But, if there is application release for same package name, the application adding should be blocked. Therefore verify if there is an application release for same package name before adding an application.