Fix issue with reconfiguring policy
The cause of the issue is not passing the newly introduced parameter 'policyType' in the policy edit payload. This parameter is marked @NotNull. Hence throwing an error. 

This fix sets the policyType parameter from the existing payload (Doesn't allow to edit as there is no UI interface to edit this parameter ATM

Fixes: entgra/product-iots#175